+++ features/encrypt/core/src/main/java/org/apache/shardingsphere/encrypt/rewrite/condition/EncryptConditionEngine.java
@@ -72,6 +72,7 @@ public final class EncryptConditionEngine {
         SUPPORTED_COMPARE_OPERATORS.add("<=");
         SUPPORTED_COMPARE_OPERATORS.add("IS");
         SUPPORTED_COMPARE_OPERATORS.add("LIKE");
+        SUPPORTED_COMPARE_OPERATORS.add("NOT LIKE");
     }
     
     /**